requirements.txt
psycopg2-binary
python-dotenv
stanza
spacy
beautifulsoup4
nltk
scikit-learn
feedparser
requests
pandas
streamlit
matplotlib
seaborn
